// frontend/src/api/endpoints/generated.ts
export type FeatEntryPrerequisites = { [key: string]: unknown };

export interface FeatEntry {
  id?: string;
  name?: string;
  description?: string;
  prerequisites?: FeatEntryPrerequisites;
}

export type ContentResponseFeats = {[key: string]: FeatEntry};

export interface SubClassEntry {
  id?: string;
  name?: string;
  description?: string;
}

export interface FeatureDef {
  level?: number;
  id?: string;
  name?: string;
}

export type SpellcastingEntrySpellSlots = {[key: string]: number};

export type SpellcastingEntrySpellLists = {[key: string]: string[]};

export interface SpellcastingEntry {
  type?: string;
  ability?: string;
  cantripsKnown?: number[];
  preparedSpells?: number[];
  spellsKnown?: number[];
  spellSlots?: SpellcastingEntrySpellSlots;
  ritualCasting?: boolean;
  spellLists?: SpellcastingEntrySpellLists;
}

export interface ClassEntry {
  id?: string;
  name?: string;
  hitDie?: string;
  savingThrows?: string[];
  primaryAbility?: string[];
  skillChoices?: number;
  skillPool?: string[];
  spellcaster?: boolean;
  subClasses?: SubClassEntry[];
  subclassLevel?: number;
  features?: FeatureDef[];
  spellcasting?: SpellcastingEntry;
}

export interface BackgroundEntry {
  id?: string;
  name?: string;
  skills?: string[];
  feat?: string;
}

export interface VariantEntry {
  id?: string;
  name?: string;
}

export interface SpeciesEntry {
  id?: string;
  name?: string;
  size?: string;
  speed?: number;
  variants?: VariantEntry[];
}

export interface AbilityEntry {
  id?: string;
  name?: string;
}

export interface SkillEntry {
  id?: string;
  name?: string;
  ability?: string;
}

export interface ContentResponse {
  classes?: ClassEntry[];
  backgrounds?: BackgroundEntry[];
  species?: SpeciesEntry[];
  abilities?: AbilityEntry[];
  skills?: SkillEntry[];
  feats?: ContentResponseFeats;
  spellCasterClasses?: string[];
}